Fix encoding problem w/passwords w/special chars
Keychain outputs passwords with non-ascii or backslashes as a strings
with octal embedded, e.g. pass\ -> pass\134. When this is parsed back
into node it doesn't pick up this embedding and it causes the password
to contain the raw octal conversion, e.g. pass\\134.

Luckily whenever Keychain does this octal escaping it also outputs a hex
version, so we can parse that entire string as UTF-8 instead.